Don't leak sessions when changing session id.
Previously COMMIT was generating a new id for the session but not
communicating it to the caller, so when FINALIZE-STATE was called it
resent the old id to the client and the session was lost (the leak).

Instead, generate the (possibly) new id in FINALIZE and pass it to
COMMIT and FINALIZE-STATE, which will send the new id to the client.